refactor: CommunityPost 타입 파일로 분리 #1478
